whale-monorepo/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.1
redis==5.2.1
celery==5.4.0
stripe==11.1.0
python-telegram-bot==21.10
pydantic==2.10.4
httpx==0.27.2
pytest==8.3.4
